+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - Fixed Gamma stuck after Final Egg.
 - Fixed Amy stuck after Hedgehog Hammer.
 - Fixed Chao garden's warp gates not working.
+- Windy Valley now requires the wind stone again.
 
 ## [v0.2.1] - 2024-07-13